refactor coupon models to separate DTOs and entities, update timestamp handling to epoch seconds, and implement social login support

+++ app/src/main/java/com/thanh0x/coursedeals/ui/binding_adapter/CourseDealBindingAdapterJava.java
@@ -44,7 +44,7 @@ public class CourseDealBindingAdapterJava {
     }
 
     @BindingAdapter("timeLeft")
-    public static void timeLeft(TextView textView, Double expiredDate) {
+    public static void timeLeft(TextView textView, Long expiredDate) {
         textView.setText(new MapperToView(textView.getContext()).mapTimeLeft(expiredDate));
     }
 }
